Add docker support
It provides a script with command line arguments for creating a
docker machine, building a docker image and running the docker
machine. The docker configurations are created based off of some
provided templates in the docker folder. These serve as the base
for creating a Dockerfile and docker-compose.yaml files. These
created files are ignored by git because they get overwritten
when the script is run.

The script enforces it's run from the root folder of the repository.
This ensures the script reads the necessary template files from the
right location. It uses the `src` folder as means of knowing that
it's being run from the root of the repo. Hopefully there's a better
way of doing this.

Update the README file with info on running the script.
Adds an entry to `.gitignore` to exclude the created files by the
script.